requirements.txt
#Librerias necesarias
fastapi
uvicorn
sqlalchemy
psycopg2-binary  # Driver para PostgreSQL
pydantic
sqlitecloud
passlib[bcrypt] 
python-jose[cryptography] 
python-multipart